Copy events.
