+++ crates/core/runner/src/runner_OLD.rs
@@ -27,187 +27,6 @@ const LOG_TARGET: &str = "moon:runner";
 
 pub type ActionResults = Vec<Action>;
 
-fn extract_run_error<T>(result: &Result<T, RunnerError>) -> Option<String> {
-    match result {
-        Ok(_) => None,
-        Err(error) => Some(error.to_string()),
-    }
-}
-
-async fn run_action(
-    node: &ActionNode,
-    action: &mut Action,
-    context: Arc<RwLock<RunnerContext>>,
-    workspace: Arc<RwLock<Workspace>>,
-    project_graph: Arc<RwLock<ProjectGraph>>,
-    emitter: Arc<RwLock<Emitter>>,
-) -> Result<(), RunnerError> {
-    let local_emitter = Arc::clone(&emitter);
-    let local_emitter = local_emitter.read().await;
-
-    let local_project_graph = Arc::clone(&project_graph);
-    let local_project_graph = local_project_graph.read().await;
-
-    let result = match node {
-        // Install dependencies in the workspace root
-        ActionNode::InstallDeps(runtime) => {
-            local_emitter
-                .emit(Event::DependenciesInstalling {
-                    project: None,
-                    runtime,
-                })
-                .await?;
-
-            let install_result = match runtime {
-                // Runtime::Node(_) => {
-                //     node_actions::install_deps(action, context, workspace, runtime, None)
-                //         .await
-                //         .map_err(RunnerError::Workspace)
-                // }
-                _ => Ok(ActionStatus::Passed),
-            };
-
-            local_emitter
-                .emit(Event::DependenciesInstalled {
-                    error: extract_run_error(&install_result),
-                    project: None,
-                    runtime,
-                })
-                .await?;
-
-            install_result
-        }
-
-        // Install dependencies in the project root
-        ActionNode::InstallProjectDeps(runtime, project_id) => {
-            let project = local_project_graph.get(project_id)?;
-
-            local_emitter
-                .emit(Event::DependenciesInstalling {
-                    project: Some(project),
-                    runtime,
-                })
-                .await?;
-
-            let install_result = match runtime {
-                // Runtime::Node(_) => {
-                //     node_actions::install_deps(action, context, workspace, runtime, Some(project))
-                //         .await
-                //         .map_err(RunnerError::Workspace)
-                // }
-                _ => Ok(ActionStatus::Passed),
-            };
-
-            local_emitter
-                .emit(Event::DependenciesInstalled {
-                    error: extract_run_error(&install_result),
-                    project: Some(project),
-                    runtime,
-                })
-                .await?;
-
-            install_result
-        }
-
-        // Run a task within a project
-        ActionNode::RunTarget(target_id) => {
-            let target = Target::parse(target_id)?;
-
-            local_emitter
-                .emit(Event::TargetRunning { target: &target })
-                .await?;
-
-            let run_result = actions::run_target(
-                action,
-                context,
-                workspace,
-                project_graph,
-                Arc::clone(&emitter),
-                &target,
-            )
-            .await;
-
-            local_emitter
-                .emit(Event::TargetRan {
-                    error: extract_run_error(&run_result),
-                    target: &target,
-                })
-                .await?;
-
-            run_result
-        }
-
-        // Setup and install the specific tool
-        ActionNode::SetupTool(runtime) => {
-            local_emitter
-                .emit(Event::ToolInstalling { runtime })
-                .await?;
-
-            let tool_result = match runtime {
-                // Runtime::Node(_) => node_actions::setup_tool(action, context, workspace, runtime)
-                //     .await
-                //     .map_err(RunnerError::Workspace),
-                _ => Ok(ActionStatus::Skipped),
-            };
-
-            local_emitter
-                .emit(Event::ToolInstalled {
-                    error: extract_run_error(&tool_result),
-                    runtime,
-                })
-                .await?;
-
-            tool_result
-        }
-
-        // Sync a project within the graph
-        ActionNode::SyncProject(runtime, project_id) => {
-            let project = local_project_graph.get(project_id)?;
-
-            local_emitter
-                .emit(Event::ProjectSyncing { project, runtime })
-                .await?;
-
-            let sync_result = match runtime {
-                // Runtime::Node(_) => {
-                //     node_actions::sync_project(action, context, workspace, project_graph, project)
-                //         .await
-                //         .map_err(RunnerError::Workspace)
-                // }
-                _ => Ok(ActionStatus::Passed),
-            };
-
-            local_emitter
-                .emit(Event::ProjectSynced {
-                    error: extract_run_error(&sync_result),
-                    project,
-                    runtime,
-                })
-                .await?;
-
-            sync_result
-        }
-    };
-
-    match result {
-        Ok(status) => {
-            action.done(status);
-        }
-        Err(error) => {
-            action.fail(error.to_string());
-
-            // If these fail, we should abort instead of trying to continue
-            if matches!(node, ActionNode::SetupTool(_))
-                || matches!(node, ActionNode::InstallDeps(_))
-            {
-                action.abort();
-            }
-        }
-    }
-
-    Ok(())
-}
-
 pub struct Runner {
     bail: bool,
